Fixed plugin to run in new model

+++ install-files/usr/share/n4d/python-plugins/ShutdownerClient.py
@@ -15,23 +15,44 @@ class ShutdownerClient:
 		self.core=n4dcore.Core.get_core()
 		self.cron_file="/etc/cron.d/lliurex-shutdowner"
 		self.desktop_cron_file="/etc/cron.d/lliurex-shutdowner-thinclients"
-		self.override_folder="/etc/lliurex-shutdowner"
-		self.override_token=os.path.join(self.override_folder,"client-override.token")
+		self.override_shutdown_folder="/etc/lliurex-shutdowner"
+		self.override_shutdown_token=os.path.join(self.override_shutdown_folder,"client-override_shutdown.token")
 		self.adi_client="/usr/bin/natfree-client"
+		self.adi_server="/usr/bin/natfree-server"
+		self.is_desktop=False
 
 	#def init
 	
 	def startup(self,options):
 		
-		if self._is_client_mode():
-			if os.path.exists(self.desktop_cron_file):
-				os.remove(self.desktop_cron_file)
-			t=threading.Thread(target=self._startup)
-			t.daemon=True
-			t.start()
+		t=threading.Thread(target=self._check_connection)
+		t.daemon=True
+		t.start()
 		
 	#def startup
 	
+	def _check_connection(self):
+		
+		if self._is_client_mode():
+			remove_cron=True
+			if self.is_desktop:
+				max_retry=602
+				count=0
+				while True:
+					if count>=max_retry:
+						break
+					count+=1
+					time.sleep(1)
+				
+				if not self._check_connection_with_server():
+					remove_cron=False
+			
+			if remove_cron:
+				if os.path.exists(self.desktop_cron_file):
+					os.remove(self.desktop_cron_file)
+			
+				self._startup()
+		
 	def _startup(self):
 		
 		self.core.register_variable_trigger("SHUTDOWNER","ShutdownerClient",self.shutdowner_trigger)
@@ -39,9 +60,7 @@ class ShutdownerClient:
 		# Making sure we're able to read SHUTDOWNER var from server
 		tries=10
 		for x in range(0,tries):
-		
 			self.shutdowner_var=self.core.get_variable("SHUTDOWNER")["return"]
-			
 			if self.shutdowner_var != None:
 				self.shutdowner_trigger(self.shutdowner_var)
 				break
@@ -56,15 +75,15 @@ class ShutdownerClient:
 	
 	def shutdowner_trigger(self,value):
 		
-		override_shut=False
-		override_enabled=self._is_override_enabled()
+		override_shutdown=False
+		override_shutdown_enabled=self._is_override_shutdown_enabled()
 
 		if value!=None:
 			if not value["cron_values"]["server_shutdown"]:
-				if override_enabled:
-					override_shut=True
+				if override_shutdown_enabled:
+					override_shutdown=True
 
-			if not override_shut:
+			if not override_shutdown:
 				if value["cron_enabled"]:
 					if value["cron_content"]!=None:
 						value["cron_content"]=value["cron_content"].replace("&gt;&gt;",">>")
@@ -87,7 +106,7 @@ class ShutdownerClient:
 
 	def is_shutdown_override_enabled(self):
 
-		is_enabled=self._is_override_enabled()
+		is_enabled=self._is_override_shutdown_enabled()
 		
 		return n4d.responses.build_successful_call_response(is_enabled)
 
@@ -98,7 +117,7 @@ class ShutdownerClient:
 		value=self.core.get_variable("SHUTDOWNER")["return"]
 		ret=False
 		if not value["cron_values"]["server_shutdown"]:
-			self._create_override_token()
+			self._create_override_shutdown_token()
 			self._delete_cron_file()
 			ret=True
 		
@@ -108,7 +127,7 @@ class ShutdownerClient:
 
 	def disable_override_shutdown(self):
 
-		self._delete_override_token()
+		self._delete_override_shutdown_token()
 		value=self.core.get_variable("SHUTDOWNER")["return"]
 
 		if value!=None:
@@ -122,17 +141,17 @@ class ShutdownerClient:
 
 	#def disable_override_shutdown
 
-	def _is_override_enabled(self):
+	def _is_override_shutdown_enabled(self):
 
-		if not os.path.exists(self.override_folder):
+		if not os.path.exists(self.override_shutdown_folder):
 			return False
 		else:
-			if os.path.exists(self.override_token):
+			if os.path.exists(self.override_shutdown_token):
 				return True
 			else:
 				return False
 
-	#def _is_override_enabled 
+	#def _is_override_shutdown_enabled 
 
 	def _create_cron_file(self,value):
 
@@ -149,28 +168,28 @@ class ShutdownerClient:
 
 	#def _delete_cron_file
 
-	def _create_override_token(self):
+	def _create_override_shutdown_token(self):
 
-		if not os.path.exists(self.override_folder):
-			os.mkdir(self.override_folder)
+		if not os.path.exists(self.override_shutdown_folder):
+			os.mkdir(self.override_shutdown_folder)
 
-		if not os.path.exists(self.override_token):
-			f=open(self.override_token,'w')
+		if not os.path.exists(self.override_shutdown_token):
+			f=open(self.override_shutdown_token,'w')
 			f.close()
 
-	#def _create_override_token
+	#def _create_override_shutdown_token
 
-	def _delete_override_token(self):
+	def _delete_override_shutdown_token(self):
 
-		if os.path.exists(self.override_token):
-			os.remove(self.override_token)
+		if os.path.exists(self.override_shutdown_token):
+			os.remove(self.override_shutdown_token)
 
-	#def _delete_override_token
+	#def _delete_override_shutdown_token
 	
 	def _is_client_mode(self):
 
 		is_client=False
-		is_desktop=False
+		self.is_desktop=False
 	
 		try:
 			cmd='lliurex-version -v'
@@ -189,15 +208,11 @@ class ShutdownerClient:
 				elif 'client' in item:
 					is_client=True
 				elif 'desktop' in item:
-					is_desktop=True
-					if os.path.exists(self.adi_client):
-						is_client=True
-			
-			if is_client:
-				if is_desktop:
-					if not self._check_connection_with_server():
-						is_client=False
-			
+					self.is_desktop=True
+					if not os.path.exists(self.adi_server):
+						if os.path.exists(self.adi_client):
+							is_client=True
+
 			return is_client
 			
 		except Exception as e:
